Break Clapi definition types out of Valuespace
because Valuespace is fat enough already!
