+++ src/Clapi/Types.hs
@@ -9,7 +9,9 @@ module Clapi.Types
 import Prelude hiding (fail)
 import Control.Monad.Fail (MonadFail, fail)
 
+import Clapi.Types.AssocList as X
 import Clapi.Types.Base as X
+import Clapi.Types.Definitions as X
 import Clapi.Types.UniqList as X
 import Clapi.Types.Wire as X
 import Clapi.Types.Messages as X
