+++ test/RelaySpec.hs
@@ -12,8 +12,8 @@ import Clapi.Path (pattern Root, toText)
 import Clapi.PathQ
 import Clapi.Types (
     DataUpdateMessage(..), TreeUpdateMessage(..), OwnerUpdateMessage(..),
-    Interpolation(IConstant), Time(..), WireValue(..))
-import Clapi.Valuespace (getType, baseValuespace, Liberty(..))
+    Interpolation(IConstant), Time(..), WireValue(..), Liberty(..))
+import Clapi.Valuespace (getType, baseValuespace)
 import Clapi.Relay (handleMessages)
 import Clapi.NamespaceTracker (Request(..), Response(..))
